test(canon): gate TS7 ProjectSession-only probes
